requirements.txt
streamlit
torch
torchvision
facenet-pytorch
opencv-python-headless
scikit-learn
Pillow
